Merge existing state into master (#3)
* vpc wip

* WIP: VPC-ify RDS and Lambda

Lambda not quite working properly yet, stupid serverless

* wip

* persist message to db

* endpoint

* Fix CFN template

* socketio logging

* POWER MOCKITO test

* Fix sends to SocketIO server

The sends to the socket IO server were being processed asynchronously
and the lambda was probably ending before they ever got sent. This adds
a sleep while waiting the socket to connect, allowing the messages to
flow through.

In addition, PowerMock was interferring with an internal library used by
SocketIO, so this fixes the unit test.

* Fix Lambda's SocketIO URL

It was missing the http:// prefix. At some point, we should factor out
the fact that it's HTTP, as it's spread across the ALB listener and the
URL, and just pass that in. But probably good enough for now.

* start adding a non-lambda handler for use from a container