Fix SocketIO server URL configuration (#5)
Due to https://github.com/facebook/create-react-app/issues/102
environment variables must be prefixed with REACT_APP_ to be made
available to react apps. The old environment variable name didn't have
that prefix and so it wasn't being configured properly. This fixes the
environment variable name so it is configured properly.